+++ include/LLVM_interface/misc.h
@@ -11,6 +11,16 @@ namespace panther{
 		
 		struct ParamInfo{
 			std::string_view name;
+			bool readonly = false;
+			bool nonnull = false;
+			bool noalias = false;
+
+			struct Dereferenceable{
+				uint64_t bytes = 0; // size{0} means can't dereferenced
+				bool can_be_null = true;
+			} dereferenceable = {};
+
+			// TODO: align?
 		};
 
 		auto setupFuncParams(llvm::Function* func, evo::ArrayProxy<ParamInfo> param_infos) noexcept -> void;